requirements.txt
aiohttp==3.7.4.post0
aiomysql==0.0.22
aiopg==1.3.3
discord-components==2.1.2
discord-pretty-help==1.3.3
